# db/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Fog of War API",
    "description": "API docs for endpoints with Fog of War app",
    "version": "0.1"
  },
  "components": {
    "securitySchemes": {
      "passport": {
        "type": "oauth2",
        "description": "Passport Authentication",
        "flows": {
          "implicit": {
            "authorizationUrl": "/auth/google",
            "scopes": {
              "read": "Grants read-only access to user profiles",
              "write": "Grants write access to user profiles"
            }
          }
        }
      }
    }
  },
  "host": "https://fog-of-war.onrender.com",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
...
    "/auth/signup": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Creates a new user account",
        "description": "# Creates a new user account with the given credentials",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "display_name": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User account created successfully"
          },
          "400": {
            "description": "Invalid input parameters"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Logs in a user",
        "description": "# Signs in a user with the given credentials",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User signed in successfully"
          },
          "400": {
            "description": "Invalid input parameters"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "passport-local": []
          }
        ]
      }
    }
  },
...
}